Utils extraction for oauth and portal improvements
